Add filesystem error avoidance
Until now there was a big TODO on filesystem errors. The code was
just blocking on an infinite loop. Now the filesystem errors try
to be handled or ignored.

This patch also does a necessary refactor on the previous class
"log_files.hpp" which had very poor naming (coding tired?).